requirements.txt
streamlit
google-generativeai==0.8.6
pandas
edge-tts
plotly